Bug fix to include TPP local codes in FHIR->enterprise and FHIR->subscriber transforms

+++ src/main/java/org/endeavourhealth/transform/emis/csv/transforms/coding/ClinicalCodeTransformer.java
@@ -78,7 +78,7 @@ public abstract class ClinicalCodeTransformer {
 
         EmisCsvCodeMap mapping = new EmisCsvCodeMap();
         mapping.setMedication(false);
-        mapping.setCodeId(codeId.getLong());
+        mapping.setCodeId(codeId.getLong().longValue());
         mapping.setCodeType(codeType.getValue());
         mapping.setReadTerm(emisTerm.getString());
         mapping.setReadCode(emisCode.getString());
